Fix TestableContinuousTrajectory friend declaration
First Clang error was that the declared typename shadows that of the
containing class. If the typename is changed, continuous_trajectory_test
fails to compile, due to access of the private variable degree_, even
though it's supposedly a friend of ContinousTrajectory. Changing the
friend class declaration to

    friend class TestableContinuousTrajectory<Frame>;

also doesn't work, with Clang complaining that this is a specialization
of a non-template class. Adding a forward declaration of
TestableContinuousTrajectory allows the code to compile.