Load Unity split DLLs only on non-Windows systems
The extra references cause errors in MSBuild, which apparently cannot be
silenced. This causes the Windows CI to fail, since warnings are treated
as errors there.

Unfortunately, the OS type is either "Windows_NT" or "Unix", and Linux
doesn't have the split DLLs. Warnings are not treated as errors for the
Linux CI, though, so it's all good.